reorg extract strings and black format
