+++ floss/stackstrings.py
@@ -11,8 +11,8 @@ import viv_utils.emulator_drivers
 
 import floss.utils
 import floss.strings
+from floss.utils import extract_strings
 from floss.results import StackString
-from floss.string_decoder import extract_strings
 
 logger = floss.logging.getLogger(__name__)
 MAX_STACK_SIZE = 0x10000
@@ -171,9 +171,7 @@ def extract_stackstrings(vw, selected_functions, min_length, quiet=False):
                 )
                 for s in extract_strings(ctx.stack_memory, min_length, seen):
                     frame_offset = (ctx.init_sp - ctx.sp) - s.offset - getPointerSize(vw)
-                    ss = StackString(
-                        fva, s.string, s.encoding, ctx.pc, ctx.sp, ctx.init_sp, s.offset, frame_offset
-                    )
+                    ss = StackString(fva, s.string, s.encoding, ctx.pc, ctx.sp, ctx.init_sp, s.offset, frame_offset)
                     # TODO option/format to log quiet and regular, this is verbose output here currently
                     logger.info("%s [%s] in 0x%x at frame offset 0x%x", ss.string, s.encoding, fva, ss.frame_offset)
                     seen.add(s.string)
